refactor(types): extract SpamCheckResult and SpamDetectionInput to shared type
- add src/lib/types/spam.ts with SpamCheckResult and SpamDetectionInput
- remove duplicate local SpamCheckResult interfaces from both admin pages
- change detectSpamIndicators parameter from SightingFormValues to SpamDetectionInput
  so call sites don't need to construct a full form object
- fix emailService.ts coordinate mapping to pass null instead of 0 for missing values
- fix Handlebars template to use indicators.length instead of indicators
  to avoid showing empty indicator lists (arrays are truthy in Handlebars)
